cypress: close the test document before reloading
The test document wasn't properly closed before
reopening. This went unnoticed until the 'unload'
phase grew longer such that the new load command
basically added a view on the existing document,
instead of create a new instance altogether.

This clearly was an issue that caused timing-based
instabilities. Now we properly close and wait
until we confirm the document is closed before
opening a new instance.

Change-Id: Icda6d7870c1352cc90002ced3c74e692fa849444
Signed-off-by: Ashod Nakashian <ashod.nakashian@collabora.co.uk>